backend/requirements.txt
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
google-auth>=2.29
Pillow>=10.0
requests>=2.31
psycopg2-binary>=2.9
django-storages[s3]>=1.14
boto3>=1.34
